core/serializers.py
from rest_framework import serializers
from core.models import Question, Answer


class QuestionSerializer(serializers.ModelSerializer):
    class Meta:
        model = Question
        fields = ['id', 'text', "created_at"]

class AnswerSerializer(serializers.ModelSerializer):
    class Meta:
        model = Answer
        fields = ["id", "text", "created_at", "user_id"]
        read_only_fields = ['user_id']