requirements.txt
Django>=4.2
djangorestframework>=3.14
psycopg2-binary>=2.9
dj-database-url>=2.0
python-dotenv>=1.0